Remove openapi-snippets dep
