+++ langwatch/package.json
@@ -105,7 +105,6 @@
     "handlebars": "^4.7.8",
     "hono": "^4.7.5",
     "hono-openapi": "^0.4.6",
-    "httpsnippet": "^3.0.1",
     "ioredis": "^5.3.2",
     "jsonwebtoken": "^9.0.2",
     "lodash.clonedeep": "^4.5.0",
@@ -128,7 +127,6 @@
     "numeral": "^2.0.6",
     "ohm-js": "^17.1.0",
     "openai": "^4.20.1",
-    "openapi-snippet": "^0.14.0",
     "papaparse": "^5.4.1",
     "path-to-regexp": "^6.3.0",
     "pino": "^9.6.0",
