+++ langwatch/src/prompt-configs/components/GeneratePromptApiSnippetButton.tsx
@@ -1,7 +1,6 @@
 import React from "react";
 import { GenerateApiSnippetButton } from "~/components/GenerateApiSnippetButton";
-import { generateGetPromptApiSnippets } from "../utils/generatePromptApiSnippet";
-import { getGetPromptSnippets } from "../utils/snippets";
+import { getGetPromptSnippets } from "../utils/snippets/getGetPromptSnippets";
 
 interface GeneratePromptApiSnippetButtonProps {
   configId: string;
@@ -13,16 +12,11 @@ interface GeneratePromptApiSnippetButtonProps {
  *
  * Renders an icon-only button that, when clicked, opens a modal (Dialog)
  * for displaying API code snippets for prompt usage.
- *
- * - SRP: This component only handles the button and modal UI.
- * - The actual code snippet generation logic will be injected later.
- * - Uses Chakra v3 and react-feather icons as per project rules.
  */
 export function GeneratePromptApiSnippetButton({
   configId,
   apiKey,
 }: GeneratePromptApiSnippetButtonProps) {
-  // const { snippets, targets } = generateGetPromptApiSnippets() ?? {};
   const snippets = getGetPromptSnippets({
     promptId: configId,
     apiKey,
